more corrections

+++ lldb/source/Plugins/Language/ObjC/NSSet.cpp
@@ -62,7 +62,7 @@ private:
 
   struct DataDescriptor_64 {
     uint64_t _used : 58;
-    uint32_t _szidx : 6;
+    uint64_t _szidx : 6;
   };
 
   struct SetItemDescriptor {
